Test : Commit all changes
