+++ k6-test/local_netty_thread_opt/k6-run-output.sh
@@ -1,0 +1,4 @@
+THREAD_CONFIG=
+k6 run \
+  --console-output=./local-thread-dump/local-netty-threaddump-$THREAD_CONFIG.log \
+  script3-thread-local-once.js
