+++ k6-test/local_netty_thread_opt/k6-run-traffic.sh
@@ -1,0 +1,1 @@
+k6 run script3-thread-local-traffic.js
